pyproject.toml
[project]
name = "csmcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.12.0",
    "fastmcp>=2.13.3",
    "fpdf2>=2.8.5",
    "isort>=7.0.0",
    "pypdf>=6.4.1",
    "pyyaml>=6.0.3",
    "reportlab>=4.4.6",
    "ruff>=0.14.8",
]
